stock-ai-analyzer/backend/requirements.txt
fastapi
uvicorn
yfinance
feedparser
google-generativeai
pandas
numpy
python-dotenv
requests